delta-rs/requirements.txt
delta-spark==2.3.0
